# docs/deapi-mirror/articles/openapi.hono.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dryAPI Gateway API",
    "description": "Unified OpenAI-compatible inference gateway with async job orchestration, provider controls, and webhook tooling.",
    "version": "2026-03-15"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin worker URL"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Use `Authorization: Bearer <api-key>`."
      }
    }
  },
  "paths": {
...
    "/v1/jobs/{surface}/{jobId}/download": {
      "get": {
        "operationId": "downloadJobResult",
        "tags": [
          "Jobs"
        ],
        "summary": "Download links or serialized payload for a completed job",
        "description": "Fetches provider result links when available, or returns a serialized payload as JSON/TXT attachment fallback.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "surface",
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "const": "chat"
                },
                {
                  "type": "string",
                  "const": "images"
                },
                {
                  "type": "string",
                  "const": "embeddings"
                },
                {
                  "type": "string",
                  "const": "transcribe"
                }
              ],
              "examples": [
                "chat"
              ]
            },
            "required": true,
            "description": "Inference surface used to select the provider endpoint."
          },
          {
            "in": "path",
            "name": "jobId",
            "schema": {
              "type": "string",
              "minLength": 1,
              "examples": [
                "f3de27f8-61d5-4d58-aad1-a7d63f8a6e0f"
              ]
            },
            "required": true,
            "description": "Provider job identifier returned from an enqueue route."
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "const": "json"
                },
                {
                  "type": "string",
                  "const": "txt"
                }
              ],
              "examples": [
                "json"
              ]
            },
            "description": "Download file format when no direct media links are available."
          }
        ],
        "responses": {
          "200": {
            "description": "Result links or serialized payload export.",
            "content": {
              "application/json": {
                "examples": {
                  "links": {
                    "summary": "Provider result links",
                    "value": {
                      "job_id": "f3de27f8-61d5-4d58-aad1-a7d63f8a6e0f",
                      "surface": "images",
                      "mode": "links",
                      "links": [
                        "https://cdn.example.com/out/abc.png"
                      ]
                    }
                  }
                }
              },
              "text/plain": {
                "example": "{\"status\":\"COMPLETED\"}"
              }
            }
          },
          "400": {
            "description": "Invalid path/query parameters or unresolved endpoint mapping."
          },
          "401": {
            "description": "Missing or invalid bearer API key."
          },
          "404": {
            "description": "Job not found for the selected surface/endpoint."
          },
          "429": {
            "description": "Rate-limited by gateway quota controls."
          },
          "500": {
            "description": "Gateway configuration or execution error."
          },
          "502": {
            "description": "Provider returned a non-success response."
          }
        }
      }
    },
...
    "/v1/queue/batch-scaling/stream": {
      "get": {
        "operationId": "streamQueueBatchScaling",
        "tags": [
          "Queue"
        ],
        "summary": "Stream live queue batch scaling updates",
        "description": "Streams server-sent events with dynamic batch scaling payloads built from queue metrics and runtime analytics.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream that emits live queue batch scaling updates."
          }
        }
      }
    },
    "/v1/queue/batch-scaling": {
      "get": {
        "operationId": "getQueueBatchScaling",
        "tags": [
          "Queue"
        ],
        "summary": "Get dynamic queue batch scaling table",
        "description": "Builds a real-time batch scaling table from queue depth snapshots and recent runtime analytics to suggest model-level batch sizes and throughput.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dynamic batch scaling table and queue summary."
          }
        }
      }
    },
...
  }
}